tried fixing rounding errors, still not working, will try using compiler flags so all calulations follow IEEE754, will also check at runtime, and if that fails, will try  only using long doubles

+++ src/hexmap/point3.hpp
@@ -287,7 +287,7 @@ public:
 class Quaternion : public Point3 {
 
 public:
-  int w;
+  double w;
   Quaternion(double x, double y, double z, double w);
 
   /**
